+++ HugeCTR/src/layers/concat_layer.cu
@@ -78,8 +78,8 @@ ConcatLayer::ConcatLayer(Tensors<float>& in_tensors, std::shared_ptr<Tensor<floa
     }
 
     int n_in_tensors = in_tensors.size();
-    int height = 0;
-    int new_width = 0;
+    size_t height = 0;
+    size_t new_width = 0;
     for (int i = 0; i < n_in_tensors; i++) {
       auto cur_in_dims = in_tensors[i]->get_dims();
       if (i != 0) {
@@ -100,7 +100,7 @@ ConcatLayer::ConcatLayer(Tensors<float>& in_tensors, std::shared_ptr<Tensor<floa
       new_width += cur_in_dims[1];
     }
 
-    std::vector<int> out_dims = {height, new_width};
+    std::vector<size_t> out_dims = {height, new_width};
     out_tensor.reset(new Tensor<float>(out_dims, blobs_buff, TensorFormat_t::HW));
 
     for (auto& in_tensor : in_tensors) {
